pyproject.toml
[project]
name = "diffing-prompts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pre-commit>=4.3.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "fire>=0.5.0",
    "hf-transfer>=0.1.9",
    "jupyter>=1.1.1",
    "datasets>=4.2.0",
    "accelerate>=1.10.1",
    "transformers==4.57.1",
    "peft==0.15.2",
    "openai>=2.3.0",
    "matplotlib>=3.10.7",
    "omegaconf>=2.3.0",
    "triton>=3.5.0",
    "bitsandbytes>=0.48.1",
    "google-genai>=1.46.0",
]
